# backend-go/docs/swagger/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/todoing.api.v1.DashboardService/GetDashboardData": {
      "post": {
        "operationId": "DashboardService_GetDashboardData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetDashboardDataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetDashboardDataRequest"
            }
          }
        ],
        "tags": [
          "DashboardService"
        ]
      }
    },
    "/todoing.api.v1.DashboardService/GetPriorityTasks": {
      "post": {
        "summary": "TaskSortConfig 复用 TaskService 的 RPC（避免重复）",
        "operationId": "DashboardService_GetPriorityTasks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPriorityTasksResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetPriorityTasksRequest"
            }
          }
        ],
        "tags": [
          "DashboardService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1DashboardData": {
      "type": "object",
      "properties": {
        "upcomingEvents": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Event"
          }
        },
        "priorityTasks": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PriorityTask"
          }
        },
        "pendingReminders": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1UpcomingReminder"
          },
          "title": "来自 reminder.proto 的 upcoming"
        },
        "eventSummary": {
          "$ref": "#/definitions/v1EventSummary"
        },
        "taskSummary": {
          "$ref": "#/definitions/v1TaskSummary"
        }
      },
      "title": "Dashboard 数据聚合"
    },
...
    "v1Event": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "eventType": {
          "$ref": "#/definitions/v1EventType"
        },
        "eventDate": {
          "type": "string",
          "format": "date-time"
        },
        "recurrenceType": {
          "$ref": "#/definitions/v1RecurrenceType",
          "title": "none 则为一次性"
        },
        "recurrenceConfig": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "预留 (例如 day, interval)"
        },
        "importanceLevel": {
          "type": "integer",
          "format": "int32",
          "description": "unchanged",
          "title": "importance_level 对齐后端 ImportanceLevel (1-5)"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "location": {
          "type": "string"
        },
        "isAllDay": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "isActive": {
          "type": "boolean"
        },
        "lastTriggeredAt": {
          "type": "string",
          "format": "date-time",
          "title": "可为空"
        }
      },
      "title": "事件"
    },
...
    "v1EventSummary": {
      "type": "object",
      "properties": {
        "totalEvents": {
          "type": "string",
          "format": "int64"
        },
        "upcomingCount": {
          "type": "string",
          "format": "int64"
        },
        "todayCount": {
          "type": "string",
          "format": "int64"
        },
        "typeBreakdown": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          }
        }
      },
      "title": "事件摘要"
    },
    "v1EventType": {
      "type": "string",
      "enum": [
        "EVENT_TYPE_UNSPECIFIED",
        "EVENT_TYPE_BIRTHDAY",
        "EVENT_TYPE_ANNIVERSARY",
        "EVENT_TYPE_HOLIDAY",
        "EVENT_TYPE_CUSTOM",
        "EVENT_TYPE_MEETING",
        "EVENT_TYPE_DEADLINE"
      ],
      "default": "EVENT_TYPE_UNSPECIFIED",
      "title": "事件类型"
    },
...
    "v1GetDashboardDataRequest": {
      "type": "object"
    },
    "v1GetDashboardDataResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        },
        "data": {
          "$ref": "#/definitions/v1DashboardData"
        }
      }
    },
...
    "v1GetPriorityTasksRequest": {
      "type": "object"
    },
    "v1GetPriorityTasksResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1Response"
        },
        "tasks": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PriorityTask"
          }
        },
        "total": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v1PriorityTask": {
      "type": "object",
      "properties": {
        "task": {
          "$ref": "#/definitions/v1Task"
        },
        "priorityScore": {
          "type": "number",
          "format": "double"
        },
        "daysLeft": {
          "type": "integer",
          "format": "int32"
        },
        "isUrgent": {
          "type": "boolean"
        }
      },
      "title": "优先任务 (Dashboard, Unified 使用)"
    },
    "v1RecurrenceType": {
      "type": "string",
      "enum": [
        "RECURRENCE_TYPE_UNSPECIFIED",
        "RECURRENCE_TYPE_NONE",
        "RECURRENCE_TYPE_YEARLY",
        "RECURRENCE_TYPE_MONTHLY",
        "RECURRENCE_TYPE_WEEKLY",
        "RECURRENCE_TYPE_DAILY"
      ],
      "default": "RECURRENCE_TYPE_UNSPECIFIED",
      "description": "- RECURRENCE_TYPE_UNSPECIFIED: 兼容",
      "title": "循环类型"
    },
...
    "v1Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/protobufAny"
        }
      },
      "title": "通用响应结构"
    },
...
    "v1Task": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/v1TaskStatus"
        },
        "priority": {
          "$ref": "#/definitions/v1TaskPriority"
        },
        "dueDate": {
          "type": "string",
          "format": "date-time",
          "title": "deprecated: 旧字段，对应后端 deadline，保留兼容"
        },
        "userId": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "scheduledDate": {
          "type": "string",
          "format": "date-time",
          "title": "scheduledDate"
        },
        "deadline": {
          "type": "string",
          "format": "date-time",
          "title": "真正截止日期"
        },
        "assignee": {
          "type": "string",
          "title": "新增: 与后端 Assignee 对齐（可能为空）"
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1TaskComment"
          },
          "title": "新增: 评论列表"
        }
      },
      "title": "任务模型"
    },
    "v1TaskComment": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "任务评论"
    },
    "v1TaskPriority": {
      "type": "string",
      "enum": [
        "TASK_PRIORITY_UNSPECIFIED",
        "TASK_PRIORITY_LOW",
        "TASK_PRIORITY_MEDIUM",
        "TASK_PRIORITY_HIGH"
      ],
      "default": "TASK_PRIORITY_UNSPECIFIED",
      "title": "任务优先级枚举"
    },
...
    "v1TaskStatus": {
      "type": "string",
      "enum": [
        "TASK_STATUS_UNSPECIFIED",
        "TASK_STATUS_TODO",
        "TASK_STATUS_IN_PROGRESS",
        "TASK_STATUS_DONE"
      ],
      "default": "TASK_STATUS_UNSPECIFIED",
      "title": "任务状态枚举"
    },
    "v1TaskSummary": {
      "type": "object",
      "properties": {
        "totalTasks": {
          "type": "string",
          "format": "int64"
        },
        "completedRate": {
          "type": "number",
          "format": "double"
        },
        "overdueCount": {
          "type": "string",
          "format": "int64"
        },
        "statusCounts": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          }
        }
      },
      "title": "任务摘要"
    },
...
    "v1UpcomingReminder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "eventTitle": {
          "type": "string"
        },
        "eventDate": {
          "type": "string",
          "format": "date-time"
        },
        "message": {
          "type": "string"
        },
        "daysLeft": {
          "type": "integer",
          "format": "int32"
        },
        "importance": {
          "type": "integer",
          "format": "int32"
        },
        "reminderAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "即将到来的提醒"
    },
...
  }
}